requirements.txt
Django==5.0.2
djangorestframework==3.14.0
google-generativeai==0.8.5
python-dotenv==1.0.1
Pillow==10.2.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
PyPDF2==3.0.1
scikit-learn==1.4.1.post1
numpy==1.26.4 